fix: add feature flag to RG - move validation to RG
